requirements.txt
pandas
flask
gunicorn
flask-restx
python-dotenv
requests
beautifulsoup4
datasets
torch
faiss-cpu
apache-airflow[celery,redis]==2.10.5
numpy
transformers==4.48.0
sentence-transformers
dvc
prefect
dvc-gs
langchain-text-splitters
langchain-community
google-cloud-storage
mlflow
langchain[mistralai]
langgraph
langfair
